Exercise 5.19 not working
I made the mistake of assuming that my code for exercise 5.19 would be
an easy fix.  I thought that if I got stuck, a fellow classmate would be
able to see where I was going wrong.  As far as Garett and I can tell,
there is nothing wrong with my code.  We looked at it for what must have
been almost two hours, and nothing came of it other than a few constants
were moved around.  I plan to have a fully working code for version 2,
but unfortunately this was all I was able to complete for version 1